Add Traefik2 and cert-manager installation
The script 'install' has been updated to include the installation steps for Traefik2 and cert-manager. This change means that k3s is set up to disable Traefik, install Traefik2 with dashboard, and also install cert-manager. Additionally, a new function, 'printer', has been added which echoes a congratulations message once the process is complete.