Setup installer script and image push action (#2)
* Add new GitHub Actions workflow for Docker build and push
* Update entrypoint and go build
* Remove pull_request trigger from goreleaser workflow
The recent changes simplify the Goreleaser GitHub workflow by removing the pull_request trigger. Now, the workflow only triggers on a push event with any tag, ensuring it only runs on version updates.
* Add tag triggers to build and publish workflow
Adjusted the CI/CD pipeline to respond not only to 'push' events on the 'main' branch, but any tagged commits as well. Removed a redundant line in the Docker build configuration.
* Create Task entries to build test multipass VM
* Remove 'version'; redundant
* Install script
* Add Traefik2 and cert-manager installation
The script 'install' has been updated to include the installation steps for Traefik2 and cert-manager. This change means that k3s is set up to disable Traefik, install Traefik2 with dashboard, and also install cert-manager. Additionally, a new function, 'printer', has been added which echoes a congratulations message once the process is complete.
* Refactor installation script for simplicity