Setup base (#4)
* Add Postgres support and separate API and Web services

* Add Alpine.js and update package.json

This commit introduces Alpine.js into the project and updates the package.json to include new scripts and necessary dependencies. The commit also generates a new lock file and modifies the index.js file for initializing Alpine.js.

* Base web server files

* Add store package and SQL queries

This commit introduces a store package with DBTX interface and functions for working with pgx. It also generates code from SQL queries using sqlc v1.26.0 tool. Addition of Dockerfile for SQL tests with postgres:15 image and sqlc.yaml configuration file are included as well.

* Add error pages and update dependencies

Added 404 and 500 error pages, along with additional structural templates to enhance the frontend experience. In addition, updates to the project's dependencies were made, including the implementation of new packages to manage the layout and improve the overall workflow.

* Update Dockerfile to build assets with PNPM and templ files

* auto generate files

* Update default PostgreSQL port in config