Add user count for pagination in ListUsers API
The commits have implemented a count of all users that an authorized user can see, which would be used for pagination. This change includes adding a countUsers function in the SQL queries and in the identity service. For the ListUsers API, the query now also returns the total count of users. Additionally, a new function PaginationQueryParams has been defined to calculate limit and offset based on page and page size.