Update user retrieval and add team membership methods
This commit modifies the existing GetUserByID method to retrieve all users in the same team as the logged-in user when given another user's ID. Additionally, two new methods, AddTeamMember and RemoveTeamMember, have been introduced to manage team memberships. However, these new methods are yet to be implemented and are currently set to throw a "not implemented" error.