Refactor and enhance identity management code
The identity management functionality has been substantially refactored and enhanced. Besides changing the way the API communicated with the database using the sqlc package, improvements have also been made in sql queries, notably in GetUserByID - which now considers teams - and in CreateTeam - which now validates if the user is an admin. Minor non-critical changes include improved comments and the addition of new helper Tasks and Docker commands for database handling.