Relocate Docker files and update paths
The Docker files have been moved to the 'zarf/docker' directory. This relocation resulted in the adjustment of paths in the Dockerfile and Taskfile.yml. Furthermore, references to docker-compose in Taskfile.yml have been updated to point to the relocated docker-compose.yml. Lastly, golines linting command has been added to Taskfile.yml.