Zarf (#8)
* Refactor code and remove unnecessary functions

* Relocate Docker files and update paths

The Docker files have been moved to the 'zarf/docker' directory. This relocation resulted in the adjustment of paths in the Dockerfile and Taskfile.yml. Furthermore, references to docker-compose in Taskfile.yml have been updated to point to the relocated docker-compose.yml. Lastly, golines linting command has been added to Taskfile.yml.

* Add Kubernetes configuration for Tawny service

This includes namespace, configmap, service, and deployment files for Tawny's API and Web servers. These files are placed under `zarf/k8s/tawny/`. Also adjusted Docker Compose and entrypoint files under `zarf/docker/` and created default values for a Postgres Helm chart.

* Refactor installation script and improve admin user handling

The installation script was refactored to remove Docker dependencies and introduce Helm for easier management. The handling of the admin user in the service initialization was changed to provide warnings instead of fatal errors when the admin user does not exist. This is to provide more flexibility and avoid crashes due to missing admin user. Notably, dependencies on Docker have been removed in favor of Kubernetes and Helm for application deployment.

* Update installation scripts to improve customizability

Made several changes to the installation scripts to allow easier customization. Specifically, exports used in the install script have been removed for better flexibility, allowing users to define their own OWNER and REPO. Helm command has been updated to handle upgrades, and helpful tools have been added to the arkade installation. The K8s location URLs were updated to be more flexible and namespaced the secret to match the repo. Also added name override option in postgres values yaml.

* Update Dockerfile path in GitHub workflow

The path to the Dockerfile in the GitHub workflow has been updated. Previously, it was not pointing to the correct location. Now it correctly points to the Dockerfile in the ./zarf/docker/ directory.

* update command

* Fix broken secret creation

* Change command to use entrypoint so that migrations run

* Update host references and add backup for etc/hosts

The commit introduces a function to backup the original /etc/hosts file before making changes. It also updates the host references in the ingress.yaml file. This is aimed to ensure smoother transition and minimize any potential conflicts with existing settings.

* Update docker build in CI